refactor(adapter): persistence 어댑터 패키지 구조 간소화
- adapter.outbound.persistence -> adapter.persistence 변경
- assessment, question, skill, user 도메인의 엔티티 및 리포지토리 패키지 이동
- import 경로 업데이트
