+++ backend/src/main/kotlin/org/setackle/backend/adapter/persistence/assessment/entity/AssessmentSessionScopeJpaEntity.kt
@@ -1,0 +1,39 @@
+package org.setackle.backend.adapter.persistence.assessment.entity
+
+import jakarta.persistence.*
+import org.setackle.backend.domain.assessment.model.AssessmentSessionScope
+
+@Entity
+@Table(name = "assessment_session_scope")
+@IdClass(AssessmentSessionScopeId::class)
+class AssessmentSessionScopeJpaEntity(
+    @Id
+    @Column(name = "session_id")
+    val sessionId: Long,
+
+    @Id
+    @Column(name = "node_id")
+    val nodeId: Long
+) {
+    fun toDomain(): AssessmentSessionScope {
+        return AssessmentSessionScope(
+            sessionId = sessionId,
+            nodeId = nodeId
+        )
+    }
+
+    companion object {
+        fun fromDomain(assessmentSessionScope: AssessmentSessionScope): AssessmentSessionScopeJpaEntity {
+            return AssessmentSessionScopeJpaEntity(
+                sessionId = assessmentSessionScope.sessionId,
+                nodeId = assessmentSessionScope.nodeId
+            )
+        }
+    }
+}
+
+@Embeddable
+data class AssessmentSessionScopeId(
+    val sessionId: Long = 0,
+    val nodeId: Long = 0
+) : java.io.Serializable
