+++ backend/src/main/kotlin/org/setackle/backend/adapter/persistence/user/entity/UserJpaEntity.kt
@@ -1,0 +1,88 @@
+package org.setackle.backend.adapter.persistence.user.entity
+
+import jakarta.persistence.*
+import org.setackle.backend.domain.user.model.*
+import org.setackle.backend.domain.user.vo.Email
+import org.setackle.backend.domain.user.vo.Password
+import org.setackle.backend.domain.user.vo.UserId
+import org.setackle.backend.domain.user.vo.UserRole
+import org.setackle.backend.domain.user.vo.Username
+import java.time.LocalDateTime
+
+@Entity
+@Table(name = "users")
+class UserJpaEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long? = null,
+
+    @Column(name = "email", nullable = false, unique = true)
+    var email: String,
+
+    @Column(name = "username", nullable = false, length = 100)
+    var username: String,
+
+    @Column(name = "password_hash", nullable = false)
+    var passwordHash: String,
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "role", nullable = false)
+    var role: UserRole = UserRole.USER,
+
+    @Column(name = "is_active", nullable = false)
+    var isActive: Boolean = true,
+
+    @Column(name = "email_verified", nullable = false)
+    var emailVerified: Boolean = false,
+
+    @Column(name = "registered_at", nullable = false, updatable = false)
+    val registeredAt: LocalDateTime = LocalDateTime.now(),
+
+    @Column(name = "updated_at", nullable = false)
+    var updatedAt: LocalDateTime = LocalDateTime.now(),
+
+    @Column(name = "last_login_at")
+    var lastLoginAt: LocalDateTime? = null,
+
+    @Column(name = "deleted_at")
+    var deletedAt: LocalDateTime? = null,
+
+    @Column(name = "deletion_reason")
+    var deletionReason: String? = null,
+) {
+    fun toDomain(): User {
+        return User.reconstruct(
+            id = id?.let { UserId.of(it) },
+            email = Email.of(email),
+            username = Username.of(username),
+            password = Password.fromHash(passwordHash),
+            role = role,
+            isActive = isActive,
+            emailVerified = emailVerified,
+            lastLoginAt = lastLoginAt
+        )
+    }
+
+    companion object {
+        fun fromDomain(user: User): UserJpaEntity {
+            return UserJpaEntity(
+                id = user.id?.value,
+                email = user.email.normalized,
+                username = user.username.normalized,
+                passwordHash = user.getHashedPassword(),
+                role = user.role,
+                isActive = user.isActive,
+                emailVerified = user.emailVerified,
+                registeredAt = LocalDateTime.now(),
+                updatedAt = LocalDateTime.now(),
+                deletedAt = null,
+                lastLoginAt = user.lastLoginAt
+            )
+        }
+    }
+
+    @PreUpdate
+    fun onPreUpdate() {
+        updatedAt = LocalDateTime.now()
+    }
+}
