+++ backend/src/main/kotlin/org/setackle/backend/adapter/web/security/JwtAuthenticationFilter.kt
@@ -4,7 +4,7 @@ import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.setackle.backend.adapter.config.JwtConfig
-import org.setackle.backend.adapter.outbound.persistence.security.CustomUserDetailsService
+import org.setackle.backend.adapter.persistence.security.CustomUserDetailsService
 import org.setackle.backend.adapter.outbound.security.JwtTokenProvider
 import org.slf4j.LoggerFactory
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
